test(research): split author-year classifier tests
